minor remove low quality images #55
